+++ CHANGES.rst
@@ -60,7 +60,11 @@ esa.euclid
   BASIC_DOWNLOAD_DATA_PRODUCTS. [#3562]
 - The method ``get_spectrum`` accepts a single source_id or designation or multiple values separated by commas or a
    list. [#3570]
-- Removed unnecessary CAST(... AS text) from array columns in ``get_product_list``, ``get_scientific_product_list`` and 
+- New method, ``get_valid_le3_configuration_values``, to retrieve the valid values for the category, group, and
+   product_type parameters dynamically. [#3601]
+- In the method, ``get_scientific_product_list``, the ``dsr_part3`` parameter now supports the additional value
+   ``latest``. [#3601]
+- Removed unnecessary CAST(... AS text) from array columns in ``get_product_list``and 
   ``__get_tile_catalogue_list``, as TAP 10.x now returns proper arrays. [#3599]